Link to GLib, use it in iutil
I cannot think of any program that would want to use hawkey but not
use librepo.  In that case, the process is already linking to GLib.

Further, libhif is a consumer of both hawkey/librepo and uses GLib (as
well as GObject/GIO, but that's a distinct topic).  It would be
significantly more pleasant for libhif if hawkey offered APIs that
used GLib primitives (such as GError and GList).

This patch merely starts the process by changing the read_whole_file()
helper function in iutil to use g_file_get_contents().  There is
quite a bit more code that could be trimmed just in iutil.

A much more significant next step would be changing the public API,
e.g. for HyPackageList be a GList<HyPackage>.  But that discussion can
come after this.