HyPackageList: Change the backing store to GPtrArray
No need to reinvent a resizable array.  One nice thing about GPtrArray
is it can hold a free function for its elements.  That means that one
can just `g_ptr_array_unref` and have the `HyPackage` elements
unreferenced too.

This doesn't change the public API yet - that would break consumers.
But this is a first step to doing that.