util: detect rpm architecture prior than machine arch
This patch implements to detect rpm architecture from
/etc/rpm/platform file. The file contains current rpm
platform architecture like "armv7hl-fedora-linux-gnu".
Originally, the file can be used to detect an arch
by rpm and yum command can also detect the architecture
through the file. Since the dnf is the default of fedora,
we can't use the file anymore.
When we want to use armv7hl platform on aarch64 kernel,
we can't use dnf command properly because the hawkey
detects the architecture by /proc/cpuinfo.

Signed-off-by: Chanho Park <chanho61.park@samsung.com>